Implement authenticate at startup.
Such that if a username and password are supplied authentication is
attempted in the ready middleware, this enables commands such as
session/get to return a session document for the user that specified the
request, ie:

rlx session set -s http://localhost:5984 -u mock-admin -p secret

Has a correct userCtx object.

Authentication information is stashed in req.login.auth so that the lazy
authentication for command invocation will not try another login if
authentication information already exists.